requirements.txt
Appium-Python-Client==2.11.1
pytesseract==0.3.10
opencv-python==4.9.0.80
langdetect==1.0.9
